Rename googlepaysheet to googlepaylauncher
