+++ payments-core/src/main/java/com/stripe/android/googlepaylauncher/GooglePayEnvironment.kt
@@ -1,4 +1,4 @@
-package com.stripe.android.googlepaysheet
+package com.stripe.android.googlepaylauncher
 
 import com.google.android.gms.wallet.WalletConstants
