+++ payments-core/src/main/java/com/stripe/android/googlepaylauncher/GooglePayLauncherConfigureViewModel.kt
@@ -1,4 +1,4 @@
-package com.stripe.android.googlepaysheet
+package com.stripe.android.googlepaylauncher
 
 import androidx.lifecycle.ViewModel
