+++ payments-core/src/main/java/com/stripe/android/googlepaylauncher/PaymentsClientFactory.kt
@@ -1,4 +1,4 @@
-package com.stripe.android.googlepaysheet
+package com.stripe.android.googlepaylauncher
 
 import android.content.Context
 import com.google.android.gms.wallet.PaymentsClient
