+++ payments-core/src/main/java/com/stripe/android/googlepaylauncher/StripeGooglePayContract.kt
@@ -1,4 +1,4 @@
-package com.stripe.android.googlepaysheet
+package com.stripe.android.googlepaylauncher
 
 import android.content.Context
 import android.content.Intent
