+++ payments-core/src/main/java/com/stripe/android/paymentsheet/DefaultGooglePayRepository.kt
@@ -7,7 +7,7 @@ import com.google.android.gms.wallet.PaymentsClient
 import com.google.android.gms.wallet.Wallet
 import com.stripe.android.GooglePayJsonFactory
 import com.stripe.android.Logger
-import com.stripe.android.googlepaysheet.GooglePayEnvironment
+import com.stripe.android.googlepaylauncher.GooglePayEnvironment
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.filterNotNull
