+++ payments-core/src/test/java/com/stripe/android/googlepaylauncher/StripeGooglePayActivityTest.kt
@@ -1,4 +1,4 @@
-package com.stripe.android.googlepaysheet
+package com.stripe.android.googlepaylauncher
 
 import android.content.Context
 import android.content.Intent
